+++ src/arduino/app_peripherals/usb_camera/__init__.py
@@ -7,7 +7,7 @@ import warnings
 from PIL import Image
 from arduino.app_peripherals.camera import Camera, CameraReadError as CRE, CameraOpenError as COE
 from arduino.app_peripherals.camera.v4l_camera import V4LCamera
-from arduino.app_utils.image.image_editor import compressed_to_png, letterboxed
+from arduino.app_utils.image import letterboxed, compressed_to_png
 from arduino.app_utils import Logger
 
 logger = Logger("USB Camera")
